Updated Insert Validate for Linux vs Docker
Updated the Insert Key, Validate Keys section to include tabs for both Linux and Docker. Specifically changed the port on the URL. Also, included an update to the script as the bool compare was failing in the Validation on lines 189 and 204, removed the quotes. This was raised in PR #79 .

calamari-bitManna

Signed-off-by: Jason2d2 <55815929+Jason2d2@users.noreply.github.com>